Move plugins to a sub-package of agent
